Extract a general preview view controller from site design preview
